Fix mapping for HiTemp devices
Fixes the mode, power and HVAC mode mappings used for HiTemp devices. This is tested
against my EvoHeat270 hot water heater.

This change is based om that of @jpcs2012 on #103, this should resolve #103.

With this change the HA Climate component is able to turn my unit on and
off and change from "intelligent" (heat in HA) to "eco" (auto in HA)
mode. The UI also now relects ON/OFF state properly as unit is turned
on/off by native app or timer.